Create Helper module. Extract setup methods in some test files to test_setup/helpers.rb
